basic_agent/requirements.txt
agent-framework[azure]>=0.1.0
azure-identity>=1.15.0
python-dotenv>=1.0.0